backend/requirements.txt
fastapi
uvicorn
yt-dlp
youtube-transcript-api
google-genai
openai
anthropic
groq
python-dotenv
sse-starlette
